rename interface files
