+++ client/src/Components/Roster/GuildMemberCard.tsx
@@ -10,8 +10,8 @@ import Typography from '@mui/material/Typography';
 import React, { useCallback, useState } from 'react';
 import { ReactComponent as DiscordLogo } from '../../assets/images/discord.svg';
 import gw2Image from '../../assets/images/gw2.png';
-import DiscordRole from '../../interfaces/DiscordRole';
-import MemberRecord from '../../interfaces/MemberRecord';
+import DiscordRole from '../../interfaces/discord-role';
+import MemberRecord from '../../interfaces/member-record';
 import { useUpdateDiscordMember } from '../../utils/apis/discord-api';
 import { useAddWarningMutation } from '../../utils/apis/warnings-api';
 import { getDateString } from '../../utils/DataProcessing';
