+++ client/src/Components/Roster/GuildMemberMenu.tsx
@@ -7,7 +7,7 @@ import Divider from '@mui/material/Divider';
 import Menu from '@mui/material/Menu';
 import { PopoverPosition } from '@mui/material/Popover';
 import { useCallback, useEffect, useState } from 'react';
-import MemberRecord from '../../interfaces/MemberRecord';
+import MemberRecord from '../../interfaces/member-record';
 import { useAdminRoles } from '../../utils/apis/auth-api';
 import GuildMemberMenuItem from './GuildMemberMenuItem';
