+++ client/src/Components/Roster/Warnings/WarningsViewer.tsx
@@ -9,8 +9,8 @@ import DialogTitle from '@mui/material/DialogTitle';
 import IconButton from '@mui/material/IconButton';
 import Typography from '@mui/material/Typography';
 import { useCallback } from 'react';
-import MemberRecord from '../../../interfaces/MemberRecord';
-import Warning from '../../../interfaces/Warning';
+import MemberRecord from '../../../interfaces/member-record';
+import Warning from '../../../interfaces/warning';
 import { useDeleteWarningMutation } from '../../../utils/apis/warnings-api';
 import useConfirm from '../../Common/ConfirmDialog/useConfirm';
 import './WarningsViewer.scss';
