+++ client/src/interfaces/auth-info.ts
