+++ client/src/interfaces/discord-log-string-generator.ts
