+++ client/src/interfaces/event-settings.ts
