+++ client/src/interfaces/event.ts
