+++ client/src/interfaces/gw2-log-entry.ts
