+++ client/src/interfaces/gw2-rank.ts
