+++ client/src/interfaces/recruitment-post.ts
