+++ client/src/utils/DiscordLogStringGenerators/MemberRoleUpdate.ts
@@ -1,8 +1,8 @@
-import { DiscordLogChange, DiscordLogUser } from '../../interfaces/DiscordLog';
+import { DiscordLogChange, DiscordLogUser } from '../../interfaces/discord-log';
 import {
   DiscordLogDisplay,
   DiscordLogDisplayGenerator
-} from '../../interfaces/DiscordLogStringGenerator';
+} from '../../interfaces/discord-log-string-generator';
 
 export class MemberRoleUpdate implements DiscordLogDisplayGenerator {
   constructor(
